requirements.txt
# Data handling
pandas==2.1.0
numpy==1.26.0
newspaper4k
requests
lxml_html_clean

# Machine Learning
scikit-learn==1.3.0

# Visualization
matplotlib==3.8.0
seaborn==0.12.2

# NLP
nltk==3.9.0
spacy==3.7.0

# OCR 
pytesseract==0.3.13
Pillow==10.0.0

